toggle panel behavior in base and apparence in dark theme
to be consistent with other view based widgets
