+++ widgets/src/base.rs
@@ -974,11 +974,6 @@ live_design!{
     BareStep = <BareStep>{}
     TurtleStep = <TurtleStep>{}
     ColorPicker = <ColorPicker>{}
-
-    // Re-export the icons in case you want to override the built-in buttons with
-    // an instance of your custom button while preserving the default icons.
-    TOGGLE_PANEL_OPEN_ICON = (TOGGLE_PANEL_OPEN_ICON)
-    TOGGLE_PANEL_CLOSE_ICON = (TOGGLE_PANEL_CLOSE_ICON)
     TogglePanelBase = <TogglePanelBase>{}
     
     DesignerBase = <DesignerBase>{}
