+++ widgets/src/theme_desktop_dark.rs
@@ -3796,23 +3796,77 @@ live_design! {
         root_view = <View> {}
     }
 
-    // Re-export the icons in case you want to override the built-in buttons with
-    // an instance of your custom button while preserving the default icons.
-    TOGGLE_PANEL_OPEN_ICON = (TOGGLE_PANEL_OPEN_ICON)
-    TOGGLE_PANEL_CLOSE_ICON = (TOGGLE_PANEL_CLOSE_ICON)
+    TOGGLE_PANEL_CLOSE_ICON = dep("crate://self/resources/icons/close_left_panel.svg")
+    TOGGLE_PANEL_OPEN_ICON = dep("crate://self/resources/icons/open_left_panel.svg")
     TogglePanel = <TogglePanelBase> {
-        persistent_content = {
-            default = {
+        flow: Overlay,
+        width: 300,
+        height: Fill,
+
+        open_content = <CachedView> {
+            width: Fill
+            height: Fill
+
+            draw_bg: {
+                instance opacity: 1.0
+    
+                fn pixel(self) -> vec4 {
+                    let color = sample2d_rt(self.image, self.pos * self.scale + self.shift) + vec4(self.marked, 0.0, 0.0, 0.0);
+                    return Pal::premul(vec4(color.xyz, color.w * self.opacity))
+                }
+            }
+        }
+
+        persistent_content = <View> {
+            height: Fit
+            width: Fill
+            default = <View> {
+                height: Fit,
+                width: Fill,
+                padding: {top: 58, left: 15, right: 15}
+                spacing: 10,
+
+                before = <View> {
+                    height: Fit,
+                    width: Fit,
+                    spacing: 10,
+                }
+
+                close = <Button> {
+                    draw_icon: {
+                        svg_file: (TOGGLE_PANEL_CLOSE_ICON),
+                    }
+                }
+
                 open = <Button> {
                     visible: false,
                     draw_icon: {
-                        svg_file: (TOGGLE_PANEL_OPEN_ICON)
+                        svg_file: (TOGGLE_PANEL_OPEN_ICON),
                     }
                 }
-                close = <Button> {
-                    draw_icon: {
-                        svg_file: (TOGGLE_PANEL_CLOSE_ICON)
-                    }
+
+                after = <View> {
+                    height: Fit,
+                    width: Fit,
+                    spacing: 10,
+                }
+            }
+        }
+
+        animator: {
+            panel = {
+                default: open,
+                open = {
+                    redraw: true,
+                    from: {all: Forward {duration: 0.3}}
+                    ease: ExpDecay {d1: 0.80, d2: 0.97}
+                    apply: {animator_panel_progress: 1.0, open_content = { draw_bg: {opacity: 1.0} }}
+                }
+                close = {
+                    redraw: true,
+                    from: {all: Forward {duration: 0.3}}
+                    ease: ExpDecay {d1: 0.80, d2: 0.97}
+                    apply: {animator_panel_progress: 0.0, open_content = { draw_bg: {opacity: 0.0} }}
                 }
             }
         }
