+++ widgets/src/toggle_panel.rs
@@ -1,121 +1,7 @@
 use crate::{button::*, makepad_derive_widget::*, makepad_draw::*, view::*, widget::*};
 
 live_design! {
-    import crate::button::ButtonBase;
-    import crate::view::ViewBase;
-    import makepad_draw::shader::std::*;
-
-    TOGGLE_PANEL_CLOSE_ICON = dep("crate://self/resources/icons/close_left_panel.svg")
-    TOGGLE_PANEL_OPEN_ICON = dep("crate://self/resources/icons/open_left_panel.svg")
-
-    FadeView = <ViewBase> {
-        optimize: Texture,
-        draw_bg: {
-            instance opacity: 1.0
-
-            texture image: texture2d
-            uniform marked: float,
-            varying scale: vec2
-            varying shift: vec2
-
-            fn vertex(self) -> vec4 {
-                let dpi = self.dpi_factor;
-                let ceil_size = ceil(self.rect_size * dpi) / dpi
-                let floor_pos = floor(self.rect_pos * dpi) / dpi
-                self.scale = self.rect_size / ceil_size;
-                self.shift = (self.rect_pos - floor_pos) / ceil_size;
-                return self.clip_and_transform_vertex(self.rect_pos, self.rect_size)
-            }
-
-            fn pixel(self) -> vec4 {
-                let color = sample2d_rt(self.image, self.pos * self.scale + self.shift) + vec4(self.marked, 0.0, 0.0, 0.0);
-                return Pal::premul(vec4(color.xyz, color.w * self.opacity))
-            }
-        }
-    }
-
-    ToggleButton = <ButtonBase> {
-        width: Fit,
-        height: Fit,
-        icon_walk: {width: 20, height: 20},
-        draw_bg: {
-            instance color: #0000
-            fn pixel(self) -> vec4 {
-                return self.color;
-            }
-        }
-        icon_walk: {width: 20, height: 20},
-        draw_icon: {
-            fn get_color(self) -> vec4 {
-                return #fff;
-            }
-        }
-    }
-
-    TogglePanelBase = {{TogglePanel}} {
-        flow: Overlay,
-        width: 300,
-        height: Fill,
-
-        open_content = <FadeView> {
-            width: Fill
-            height: Fill
-        }
-
-        persistent_content = <ViewBase> {
-            height: Fit
-            width: Fill
-            default = <ViewBase> {
-                height: Fit,
-                width: Fill,
-                padding: {top: 58, left: 15, right: 15}
-                spacing: 10,
-
-                before = <ViewBase> {
-                    height: Fit,
-                    width: Fit,
-                    spacing: 10,
-                }
-
-                close = <ToggleButton> {
-                    draw_icon: {
-                        svg_file: (TOGGLE_PANEL_CLOSE_ICON),
-                    }
-                }
-
-                open = <ToggleButton> {
-                    visible: false,
-                    draw_icon: {
-                        svg_file: (TOGGLE_PANEL_OPEN_ICON),
-                    }
-                }
-
-                after = <ViewBase> {
-                    height: Fit,
-                    width: Fit,
-                    spacing: 10,
-                }
-            }
-        }
-
-        animator: {
-            panel = {
-                default: open,
-                open = {
-                    redraw: true,
-                    from: {all: Forward {duration: 0.3}}
-                    ease: ExpDecay {d1: 0.80, d2: 0.97}
-                    apply: {animator_panel_progress: 1.0, open_content = { draw_bg: {opacity: 1.0} }}
-                }
-                close = {
-                    redraw: true,
-                    from: {all: Forward {duration: 0.3}}
-                    ease: ExpDecay {d1: 0.80, d2: 0.97}
-                    apply: {animator_panel_progress: 0.0, open_content = { draw_bg: {opacity: 0.0} }}
-                }
-            }
-        }
-    }
+    TogglePanelBase = {{TogglePanel}} {}
 }
 
 /// A toggable side panel that can be expanded and collapsed to a maximum and minimum size.
